refactor(DB): 테이블 명 수정 및 테이블 이미 존재할 때만 생성하도록 수정
