+++ src/main/java/racingcar/dao/ParticipantDao.java
@@ -6,17 +6,17 @@ import org.springframework.stereotype.Repository;
 import racingcar.dto.ParticipateDto;
 
 @Repository
-public class ParticipatesDao {
+public class ParticipantDao {
 
     private final JdbcTemplate jdbcTemplate;
 
     @Autowired
-    public ParticipatesDao(final JdbcTemplate jdbcTemplate) {
+    public ParticipantDao(final JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
     public void save(final ParticipateDto participateDto) {
-        final String sql = "INSERT INTO PARTICIPATES(game_id, player_id, position, is_winner) VALUES(?, ?, ?, ?) ";
+        final String sql = "INSERT INTO PARTICIPANT(game_id, player_id, position, is_winner) VALUES(?, ?, ?, ?) ";
         jdbcTemplate.update(sql,
                 participateDto.getGameId(),
                 participateDto.getPlayerId(),
