+++ src/main/resources/data.sql
@@ -1,4 +1,4 @@
-CREATE TABLE GAME
+CREATE TABLE IF NOT EXISTS GAME
 (
     id          BIGINT   NOT NULL AUTO_INCREMENT,
     trial_count INT      NOT NULL,
@@ -6,14 +6,15 @@ CREATE TABLE GAME
     PRIMARY KEY (id)
 );
 
-CREATE TABLE PLAYER
+CREATE TABLE IF NOT EXISTS PLAYER
 (
     id   BIGINT      NOT NULL AUTO_INCREMENT,
-    name varchar(10) NOT NULL,
-    PRIMARY KEY (id)
+    name varchar(10) NOT NULL UNIQUE,
+    PRIMARY KEY (id),
+
 );
 
-CREATE TABLE PARTICIPATES
+CREATE TABLE IF NOT EXISTS PARTICIPANT
 (
     game_id   BIGINT  NOT NULL,
     player_id BIGINT  NOT NULL,
